chore: Retire HomeContainer
